+++ packages/unexus-installer/include/installerbackend.h
@@ -19,11 +19,21 @@ class InstallerBackend : public QObject
     Q_PROPERTY(bool setupAvailable READ setupAvailable NOTIFY prerequisitesChanged)
     Q_PROPERTY(bool diagnosticsAvailable READ diagnosticsAvailable NOTIFY prerequisitesChanged)
     Q_PROPERTY(bool canInstall READ canInstall NOTIFY prerequisitesChanged)
+    Q_PROPERTY(bool diskInstallAvailable READ diskInstallAvailable NOTIFY prerequisitesChanged)
+    Q_PROPERTY(bool canDiskInstall READ canDiskInstall NOTIFY diskOptionsChanged)
     Q_PROPERTY(int progress READ progress NOTIFY progressChanged)
     Q_PROPERTY(QVariantList readinessChecks READ readinessChecks NOTIFY prerequisitesChanged)
     Q_PROPERTY(QVariantList installSteps READ installSteps NOTIFY installStepsChanged)
     Q_PROPERTY(bool installGamingLaunchers READ installGamingLaunchers WRITE setInstallGamingLaunchers NOTIFY optionsChanged)
     Q_PROPERTY(bool configureBootloader READ configureBootloader WRITE setConfigureBootloader NOTIFY optionsChanged)
+    Q_PROPERTY(QString diskTarget READ diskTarget WRITE setDiskTarget NOTIFY diskOptionsChanged)
+    Q_PROPERTY(QString diskUsername READ diskUsername WRITE setDiskUsername NOTIFY diskOptionsChanged)
+    Q_PROPERTY(QString diskHostname READ diskHostname WRITE setDiskHostname NOTIFY diskOptionsChanged)
+    Q_PROPERTY(QString diskTimezone READ diskTimezone WRITE setDiskTimezone NOTIFY diskOptionsChanged)
+    Q_PROPERTY(QString diskLocale READ diskLocale WRITE setDiskLocale NOTIFY diskOptionsChanged)
+    Q_PROPERTY(QString diskKeymap READ diskKeymap WRITE setDiskKeymap NOTIFY diskOptionsChanged)
+    Q_PROPERTY(QString diskFilesystem READ diskFilesystem WRITE setDiskFilesystem NOTIFY diskOptionsChanged)
+    Q_PROPERTY(QString diskNetworkMode READ diskNetworkMode WRITE setDiskNetworkMode NOTIFY diskOptionsChanged)
     Q_PROPERTY(QString repoRoot READ repoRoot CONSTANT)
 
 public:
@@ -39,6 +49,8 @@ public:
     bool setupAvailable() const;
     bool diagnosticsAvailable() const;
     bool canInstall() const;
+    bool diskInstallAvailable() const;
+    bool canDiskInstall() const;
     int progress() const;
     QVariantList readinessChecks() const;
     QVariantList installSteps() const;
@@ -46,12 +58,30 @@ public:
     bool configureBootloader() const;
     void setInstallGamingLaunchers(bool enabled);
     void setConfigureBootloader(bool enabled);
+    QString diskTarget() const;
+    QString diskUsername() const;
+    QString diskHostname() const;
+    QString diskTimezone() const;
+    QString diskLocale() const;
+    QString diskKeymap() const;
+    QString diskFilesystem() const;
+    QString diskNetworkMode() const;
+    void setDiskTarget(const QString &target);
+    void setDiskUsername(const QString &username);
+    void setDiskHostname(const QString &hostname);
+    void setDiskTimezone(const QString &timezone);
+    void setDiskLocale(const QString &locale);
+    void setDiskKeymap(const QString &keymap);
+    void setDiskFilesystem(const QString &filesystem);
+    void setDiskNetworkMode(const QString &mode);
     QString repoRoot() const;
 
     Q_INVOKABLE void install();
     Q_INVOKABLE void repair();
     Q_INVOKABLE void diagnose();
     Q_INVOKABLE void uninstall();
+    Q_INVOKABLE void previewDiskInstall();
+    Q_INVOKABLE void installDisk();
     Q_INVOKABLE void refresh();
     Q_INVOKABLE void clearLog();
 
@@ -65,6 +95,7 @@ signals:
     void progressChanged();
     void installStepsChanged();
     void optionsChanged();
+    void diskOptionsChanged();
 
 private slots:
     void readOutput();
@@ -80,6 +111,7 @@ private:
     QString scriptPath(const QString &name) const;
     QVariantMap checkItem(const QString &label, const QString &value, const QString &status) const;
     QVariantMap stepItem(const QString &label, const QString &detail, const QString &status) const;
+    QStringList diskInstallArguments(bool execute) const;
     static bool commandExists(const QString &command);
 
     QProcess m_process;
@@ -87,6 +119,14 @@ private:
     bool m_installed = false;
     bool m_installGamingLaunchers = true;
     bool m_configureBootloader = true;
+    QString m_diskTarget;
+    QString m_diskUsername = QStringLiteral("unexus");
+    QString m_diskHostname = QStringLiteral("unexus-os");
+    QString m_diskTimezone = QStringLiteral("UTC");
+    QString m_diskLocale = QStringLiteral("en_US.UTF-8");
+    QString m_diskKeymap = QStringLiteral("us");
+    QString m_diskFilesystem = QStringLiteral("btrfs");
+    QString m_diskNetworkMode = QStringLiteral("auto");
     int m_progress = 0;
     QString m_currentAction;
     QString m_statusTitle;